+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Breaking Changes
   to ``templates/comments/form.html``.
 * Removed no longer used ``apps.blog.admin.CommentAdmin``.
 * Provided ability to disallow post comments.
+* Renamed ``branch.settings.get_emails`` to ``branch.settings._parse_emails``.
 
 
 Features